Used pthread_join() to wait for threads to finish, instead of semaphore.
Made the cache start its listener before initializing the database connections.

Fixed the logic for dividing up the work when the cache is warmed up.

Re-added the logic to parallelize the work to warm up the cache.

Added the option to define the path to write output files.

Defined a sanity check option (-s) to output the entire contents of the cache
when it warms up, and how the division of labor is done for the parallelized
warm up.

Cache requests are not serviced until the cache is warm.

Added various output messages to display the cache progress in warming up.

Applied a consistent style in various parts of the code.


git-svn-id: https://osdldbt.svn.sourceforge.net/svnroot/osdldbt/trunk/dbt1@664 eccdf8fb-1411-0410-9b6e-c06b97aefaee