Date: Tue, 21 Sep 2004 13:15:19 +1000 To: OSDL DB <osdldbt-general@lists.sourceforge.net> From: Darren Williams <dsw@gelato.unsw.edu.au> Subject: [osdldbt-general] dbt1 include config.h patch ver2
[-- Attachment #1 --]
[-- Type: text/plain, Encoding: 7bit, Size: 0.9K --]

Hi Mark

A week or so ago I sent in some patches to fix
the BK tree build for dbt1. While they work here
they require some tweeking. Could you do a

bk undo -a1.100

This should revert dbt1 back to the version prior
to all my patches being applied and apply the
attached patches, these are less intrusive to the
tree as a whole and have been tested on both
Debian and RedHat systems.
Note I have not tested an ODBC version.

The problem with the first round of pathes is that
they assume that no other config.h is in the
include search path, as I discovered on one of our
test machines this is a bad assumption.

patch
01      Incorrect condition for DB interface selection
02-04   Use correct config.h.

Sorry about the stuff up
Darren


git-svn-id: https://osdldbt.svn.sourceforge.net/svnroot/osdldbt/trunk/dbt1@1288 eccdf8fb-1411-0410-9b6e-c06b97aefaee